+++ CHANGELOG.md
@@ -6,9 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+
+### Removed
+
+- `es-verifactu-v1`: removed preceding validations for credit notes and `F3` invoice type
+
 ### Fixed
 
 - `es`: IGIC now uses VAT keys.
+
 
 ## [v0.300.1] - 2025-09-12