+++ CHANGELOG.md
@@ -10,9 +10,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - `org`: `Item` - `images` field for storing links to images of the item.
 
+### Removed
+
+- `es-verifactu-v1`: removed preceding validations for credit notes and `F3` invoice type
+
 ### Fixed
 
 - `es`: IGIC now uses VAT keys.
+
 
 ## [v0.300.1] - 2025-09-12