+++ CHANGELOG.md
@@ -8,13 +8,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 
+- `org`: `Item` - `images` field for storing links to images of the item.
 - `pt-saft-v1`: positive quantity validation
 - `pt-saft-v1`: tax rate normalization to prevent rate extension and percent mismatches
 - `pt-saft-v1`: exemption note text length validation
 
+### Removed
+
+- `es-verifactu-v1`: removed preceding validations for credit notes and `F3` invoice type
+
 ### Fixed
 
-- `es`: IGIC now uses VAT keys.
+- `es`: IGIC now uses VAT keys
+- `bill`: fixed zero-percent handling in charges and discounts
+
 
 ## [v0.300.1] - 2025-09-12